Safari-609.1.11.1/LayoutTests/crypto/subtle/ecdh-generate-export-key-spki-p384-expected.txt

Test exporting a P-384 ECDH public key with SPKI format.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key pair...
Exporting the public key...
PASS publicKey.byteLength is 120
PASS successfullyParsed is true

TEST COMPLETE
